Move start coords utils to start-coords/utils.ts
