+++ packages/perseus-editor/src/widgets/__stories__/interactive-graph-editor.stories.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 
 import {flags} from "../../__stories__/flags-for-api-options";
-import {getDefaultFigureForType} from "../interactive-graph-editor/locked-figures/util";
 import InteractiveGraphEditor from "../interactive-graph-editor/interactive-graph-editor";
+import {getDefaultFigureForType} from "../interactive-graph-editor/locked-figures/util";
 
 import InteractiveGraphEditorArgTypes from "./interactive-graph-editor.argtypes";
