+++ packages/perseus-editor/src/widgets/__tests__/interactive-graph-editor-locked-figures.test.tsx
@@ -6,8 +6,8 @@ import * as React from "react";
 
 import {testDependencies} from "../../../../../testing/test-dependencies";
 import {flags} from "../../__stories__/flags-for-api-options";
-import {getDefaultFigureForType} from "../interactive-graph-editor/locked-figures/util";
 import InteractiveGraphEditor from "../interactive-graph-editor/interactive-graph-editor";
+import {getDefaultFigureForType} from "../interactive-graph-editor/locked-figures/util";
 
 import type {PerseusGraphType} from "@khanacademy/perseus";
 import type {UserEvent} from "@testing-library/user-event";
