+++ packages/perseus-editor/src/widgets/interactive-graph-editor/components/interactive-graph-settings.test.tsx
@@ -4,6 +4,7 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
 import {testDependencies} from "../../../../../../testing/test-dependencies";
+
 import InteractiveGraphSettings from "./interactive-graph-settings";
 
 import type {UserEvent} from "@testing-library/user-event";
