+++ packages/perseus-editor/src/widgets/interactive-graph-editor/interactive-graph-editor.tsx
@@ -17,17 +17,18 @@ import * as React from "react";
 import invariant from "tiny-invariant";
 import _ from "underscore";
 
-import LabeledRow from "./locked-figures/labeled-row";
-import LockedFiguresSection from "./locked-figures/locked-figures-section";
+import InteractiveGraphDescription from "../../components/interactive-graph-description";
+import {parsePointCount} from "../../util/points";
+
 import GraphPointsCountSelector from "./components/graph-points-count-selector";
 import GraphTypeSelector from "./components/graph-type-selector";
 import {InteractiveGraphCorrectAnswer} from "./components/interactive-graph-correct-answer";
-import InteractiveGraphDescription from "../../components/interactive-graph-description";
 import InteractiveGraphSettings from "./components/interactive-graph-settings";
 import SegmentCountSelector from "./components/segment-count-selector";
+import LabeledRow from "./locked-figures/labeled-row";
+import LockedFiguresSection from "./locked-figures/locked-figures-section";
 import StartCoordsSettings from "./start-coords/start-coords-settings";
-import {shouldShowStartCoordsUI} from "../../components/util";
-import {parsePointCount} from "../../util/points";
+import {shouldShowStartCoordsUI} from "./start-coords/util";
 
 import type {
     APIOptionsWithDefaults,
