+++ packages/perseus-editor/src/widgets/interactive-graph-editor/start-coords/start-coords-settings.test.tsx
@@ -6,6 +6,7 @@ import * as React from "react";
 
 import {testDependencies} from "../../../../../../testing/test-dependencies";
 import {clone} from "../../../util/object-utils";
+
 import StartCoordsSettings from "./start-coords-settings";
 
 import type {CollinearTuple, Coord, Range} from "@khanacademy/perseus";
